draggable number text field input

+++ kool-editor-api/src/commonMain/kotlin/de/fabmax/kool/editor/api/EditorInfo.kt
@@ -3,7 +3,7 @@ package de.fabmax.kool.editor.api
 @Target(AnnotationTarget.PROPERTY)
 annotation class EditorInfo(
     val label: String = "",
-    val min: Float = Float.NEGATIVE_INFINITY,
-    val max: Float = Float.POSITIVE_INFINITY,
+    val min: Double = Double.NEGATIVE_INFINITY,
+    val max: Double = Double.POSITIVE_INFINITY,
     val hideInEditor: Boolean = false
 )
